New Styling and cached methods
feat: New Effect on CSS & Cached instances of the Date() methods

A hover effect was added onto the .mode children and the methods of Date were cached.

Conclusion: In later pushes, you should be able to change the mode in which time is displayed and the weekday should also be indicated.